wip: moving engine out of the main to be shared between codes
