+++ spooky-core/src/lib.rs
@@ -1,4 +1,5 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 pub mod assets;
+pub mod engine;
 pub mod maze;
 pub mod spritebuf;
